Fix phpspec configuration and specifications
